拆分 SFTP 前端模块并修正回归契约
将 SFTP 连接与目录管理、文本编辑器、SVG/图片预览拆分为独立脚本，降低单文件复杂度并满足前端模块边界限制。保持现有脚本加载顺序、国际化扫描和运行时行为不变。

同步终端快速命令栏检查，使其验证工作区缓存后的 terminalSurface 挂载目标。已完成构建、严格类型检查、前端与国际化检查及完整 210 项回归。
