+++ public/index.html
@@ -412,6 +412,8 @@
 <script src="/app-sftp-diff.js?v=__TERMA_VERSION__"></script>
 <script src="/app-sftp-open.js?v=__TERMA_VERSION__"></script>
 <script src="/app-sftp.js?v=__TERMA_VERSION__"></script>
+<script src="/app-sftp-editor.js?v=__TERMA_VERSION__"></script>
+<script src="/app-sftp-preview.js?v=__TERMA_VERSION__"></script>
 <script src="/app-sftp-drag-payload.js?v=__TERMA_VERSION__"></script>
 <script src="/app-sftp-drag.js?v=__TERMA_VERSION__"></script>
 <script src="/app-sftp-operations.js?v=__TERMA_VERSION__"></script>
